move op test out of queue_test.go
